Improve daily log rotation.
The previous method to use numbers 000~999 for log rotated filenames breaks if
we reach more than 999 files and when we try to compress the files as they
change name causing old logs to be rewritten.

Instead of using this naming convention this patch uses the date (2006-01-01) to
name the rotated files for daily rotation.

Maxline and Maxsize rotation methods were removed completely as these don't work
with dates as filenames.